6.10签到页面搭建

+++ quotainXM/quotainXM/Classes/Information/View/ZBInfoCollectionView.h
@@ -12,7 +12,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface ZBInfoCollectionView : UIView
 
--(instancetype)initWithW:(NSInteger )colW;
+-(instancetype)initWithW:(CGFloat )colW;
 
 @end
